clean up env_dir
