+++ test/integration/containerized/test_container_management.py
@@ -74,6 +74,10 @@ def test_cancel_will_remove_container(test_data_dir, container_runtime_installed
         cli, container_runtime_installed, 'ansible_runner_foo_bar'
     ), 'Found a running container, they should have all been stopped'
 
+    # clean up settings artifacts generated by test
+    if os.path.exists(env_dir):
+        shutil.rmtree(env_dir)
+
 
 @pytest.mark.parametrize('runtime', ['podman', 'docker'])
 def test_invalid_registry_host(tmp_path, runtime):
